Ok, food in red if out of date
So, I had to rewrite the database table to include color (this lets us
use color to show out of date items) So we should probably flush the
heroku DB info. Other little things also had to be modified for it to
all work. Gonna do one last push in a little to clean out some things.